pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ward-sdk"
version = "0.1.0"
description = "Zero-code observability for LLM applications"
readme = "readme.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "Ward Team"}
]
keywords = ["opentelemetry", "llm", "observability", "tracing", "openai", "anthropic", "genai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "opentelemetry-api>=1.20.0,<2.0.0",
    "opentelemetry-sdk>=1.20.0,<2.0.0",
    "opentelemetry-exporter-otlp-proto-http>=1.20.0,<2.0.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0,<2.0.0",
    "opentelemetry-instrumentation>=0.41b0",
    "wrapt>=1.14.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.18.0"]
all = ["anthropic>=0.18.0"]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.18.0",
]

[project.urls]
Homepage = "https://github.com/Ward-Agents-Obs/ward"
Repository = "https://github.com/Ward-Agents-Obs/ward"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["src/tests"]
asyncio_mode = "auto"
